refactor: extract createElement function into separate module
