+++ nonograms/src/js/buttons.module.js
@@ -1,4 +1,4 @@
-import { createElement } from "./page.module";
+import { createElement } from "./createElement.module";
 import { clickSaveButton } from "./save.module";
 import { clickContinueButton } from "./continue.module";
 import { clickResetButton } from "./reset.module";
