+++ nonograms/src/js/game.module.js
@@ -1,4 +1,4 @@
-import { createElement } from "./page.module";
+import { createElement } from "./createElement.module";
 import { transformAndInvertMatrix } from "./topMatrix.module";
 import { transformLeftMatrix } from "./leftMatrix.module";
 import {
