+++ nonograms/src/js/header.module.js
@@ -1,4 +1,4 @@
-import { createElement } from "./page.module";
+import { createElement } from "./createElement.module";
 import { createMenu } from "./menu.module";
 
 export let headerWrapper;
